requirements.txt
click==8.0.4
dataclasses==0.6
dominate==2.6.0
f==0.0.1
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
six==1.16.0
SQLAlchemy==1.4.27
typing_extensions==4.0.0
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
